Move cat_helper to utils/ dir.
