+++ gslib/utils/cat_helper.py
